Chronicles of the Corpse: AI’s Murder Tale
The system mainly target the law enforcement and investigators to achieve the justice. Also, the
system will be fully assessed by the investigator and law enforcement for maintenance and a high
accuracy for crime analysis. In addition, the system is designed to keep the human in the loop.
However, all users must have:
- Basic Web handling skills e.g connecting to the internet, and signing into a software.
- A brief understanding of what the forensic system is and how it works.
- A brief of working with database system and how AI works.